Restore Online status on next launch as well
Despite persisting the "Online" status to disk after the program quits, Deceive only restores offline or mobile status on next launch. This is by design, to prevent anyone from forgetting the Online status enabled and then launching the program -- knowing that they will *always* start as offline, no matter what, brings some peace of mind to the user.

Some people, however, use Deceive to stay *online* forever, due to the aggressive AFK policy of the LoL client, which sets you as "away" if you don't interact for as short as 5 minutes, causing you to go to the bottom of your friends' friend list and preventing them from inviting you into matches.

This commit allows the online status to be restored as well, which makes things a bit easier for players whose only usage of Deceive is to stay permanently online. Note that this doesn't account for any of the messages shown on startup, i.e. Deceive will still "behave" as if you were offline.